Update Recoil setting - MF 적용
