+++ packages/santa_close_frame/package.json
@@ -15,6 +15,7 @@
     "santa_close_design-system": "^1.0.0",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "recoil": "^0.7.2",
     "@karrotframe/navigator": "^0.24.2"
   },
   "devDependencies": {
