+++ packages/santa_close_frame/src/App.tsx
@@ -1,6 +1,7 @@
-import React from 'react'
+import React, {ChangeEvent} from 'react'
 import {Navigator, Screen} from '@karrotframe/navigator'
-// import {Box} from 'santa_close_design-system'
+import {RecoilRoot, useRecoilState} from 'recoil'
+import {sampleState} from 'map-app/atoms'
 
 const MapApp = React.lazy(() => import('map-app/MapApp'))
 
@@ -13,6 +14,16 @@ const Page2 = () => {
 }
 
 const MapAppContainer = () => {
+  const [state, setState] = useRecoilState<string>(sampleState)
+
+  const handleStateChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const {
+      target: {value},
+    } = e
+
+    setState(value)
+  }
+
   return (
     <div
       style={{
@@ -23,6 +34,8 @@ const MapAppContainer = () => {
       <h1>This is Frame App of santa-close-frame with KarrotFrame</h1>
       <React.Suspense fallback="loading...">
         <MapApp />
+        <input value={state} onChange={handleStateChange} />
+        <h3>{state}</h3>
       </React.Suspense>
     </div>
   )
@@ -30,11 +43,13 @@ const MapAppContainer = () => {
 
 const App = () => {
   return (
-    <Navigator onClose={console.log}>
-      <Screen component={MapAppContainer} path="/" />
-      <Screen component={Page1} path="/page1" />
-      <Screen component={Page2} path="/page2" />
-    </Navigator>
+    <RecoilRoot>
+      <Navigator onClose={console.log}>
+        <Screen component={MapAppContainer} path="/" />
+        <Screen component={Page1} path="/page1" />
+        <Screen component={Page2} path="/page2" />
+      </Navigator>
+    </RecoilRoot>
   )
 }
