+++ packages/santa_close_frame/src/index.ts
@@ -1,0 +1,1 @@
+import('./bootstrap')
