+++ packages/santa_close_map/src/MapApp.tsx
@@ -1,0 +1,32 @@
+import {ChangeEvent} from 'react'
+import {useRecoilState} from 'recoil'
+import {Box} from 'santa_close_design-system'
+import {sampleState} from './store'
+
+const MapApp = () => {
+  const [state, setState] = useRecoilState(sampleState)
+
+  const handleStateChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const {
+      target: {value},
+    } = e
+
+    setState(value)
+  }
+
+  return (
+    <div
+      style={{
+        margin: '10px',
+        border: '1px solid red',
+      }}
+    >
+      <h1>Map App with Recoil</h1>
+      <input value={state} onChange={handleStateChange} />
+      <p>echo: {state}</p>
+      <Box />
+    </div>
+  )
+}
+
+export default MapApp
