+++ packages/santa_close_map/src/index.ts
@@ -1,0 +1,1 @@
+import('./bootstrap')
