+++ packages/santa_close_map/tsconfig.json
@@ -1,15 +1,11 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "jsx": "react-jsx",
     "module": "esnext",
-    "isolatedModules": true
+    "jsx": "react-jsx",
+    "target": "ES2015"
+    // "isolatedModules": true
   },
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "build"
-  ]
+  "include": ["src/**/*"],
+  "exclude": ["node_modules"]
 }
